refactor(desktop): extract the UI-control HTTP bridge from main.mjs
The loopback control server (/health, /snapshot, /actions, /execute →
renderer window.__openworkControl via executeJavaScript; discovery file +
bearer token) moves to electron/ui-control-server.mjs as a
createUiControlServer factory (createRuntimeManager pattern). main.mjs
injects appName/appIdentifier/getWindow and owns only start/stop call
sites; the server state (instance, token, discovery path) is now private
to the module.

Tested:
- node --check both; typecheck:electron clean; check:electron 50 methods;
  desktop node:test 18 pass
- LIVE: booted the Electron dev app — discovery file written, /health ok,
  authorized /snapshot returns a real renderer snapshot (route + actions)
  through the injected window factory, unauthorized request rejected.
